+++ src/main/java/com/hellomeritz/chat/service/ChatService.java
@@ -38,76 +38,29 @@ import java.util.List;
 
 @Service
 public class ChatService {
-    private static final String SIMPLE_CIRCUIT_BREAKER_CONFIG = "simpleCircuitBreakerConfig";
-    private static final String TRANSLATION_FALLBACK_BOT_MESSAGE = "번역 관련 fallback 메서드가 호출되었습니다.";
     private static final int CHAT_PAGE_SIZE = 20;
-    private static final int MAX_ATTENDANCE_COUNT = 2;
 
     private final ChatMessageRepository chatMessageRepository;
     private final ChatRoomRepository chatRoomRepository;
-    private final TranslatorHandler translatorHandler;
-    private final AudioUploader audioUploader;
-    private final SttManagerHandler sttManagerHandler;
     private final PasswordEncoder passwordEncoder;
     private final IpSensor ipSensor;
-    private final CircuitBreakerBot circuitBreakerBot;
     private final ChatRoomEntryRepository chatRoomEntryRepository;
     private final ChatSessionRepository chatSessionRepository;
 
 
     public ChatService(
         ChatMessageRepository chatMessageRepository,
-        ChatRoomRepository chatRoomRepository, TranslatorHandler translatorHandler,
-        AudioUploader audioUploader,
-        SttManagerHandler sttManagerHandler,
+        ChatRoomRepository chatRoomRepository,
         PasswordEncoder passwordEncoder,
-        IpSensor ipSensor, CircuitBreakerBot circuitBreakerBot, ChatRoomEntryRepository chatRoomEntryRepository, ChatSessionRepository chatSessionRepository) {
+        IpSensor ipSensor,
+        ChatRoomEntryRepository chatRoomEntryRepository,
+        ChatSessionRepository chatSessionRepository) {
         this.chatMessageRepository = chatMessageRepository;
         this.chatRoomRepository = chatRoomRepository;
-        this.translatorHandler = translatorHandler;
-        this.audioUploader = audioUploader;
-        this.sttManagerHandler = sttManagerHandler;
         this.passwordEncoder = passwordEncoder;
         this.ipSensor = ipSensor;
-        this.circuitBreakerBot = circuitBreakerBot;
         this.chatRoomEntryRepository = chatRoomEntryRepository;
         this.chatSessionRepository = chatSessionRepository;
-    }
-
-    @CircuitBreaker(name = SIMPLE_CIRCUIT_BREAKER_CONFIG, fallbackMethod = "fallbackSendMessage")
-    @Transactional
-    public ChatMessageTranslateResults translateText(ChatMessageTextParam param) {
-        Translator translator = translatorHandler.getTranslator(TranslateProvider.DEEPL.name());
-        TranslationResponse translatedResponse = translator.translate(param.toTranslationRequest());
-
-        if (chatRoomEntryRepository.getAttendanceCount() == MAX_ATTENDANCE_COUNT) {
-            return ChatMessageTranslateResults.to(
-                chatMessageRepository.save(param.toReadChatMessage()),
-                chatMessageRepository.save(param.toReadChatMessage(translatedResponse.translatedText()))
-            );
-        }
-        return ChatMessageTranslateResults.to(
-            chatMessageRepository.save(param.toNotReadChatMessage()),
-            chatMessageRepository.save(param.toNotReadChatMessage(translatedResponse.translatedText()))
-        );
-    }
-
-    private ChatMessageTranslateResults fallbackSendMessage(ChatMessageTextParam param, Exception exception) {
-        circuitBreakerBot.sendBotMessage(TRANSLATION_FALLBACK_BOT_MESSAGE + ", exception: " + exception.getMessage());
-
-        Translator translator = translatorHandler.getTranslator(TranslateProvider.GOOGLE.name());
-        TranslationResponse translatedResponse = translator.translate(param.toTranslationRequest());
-
-        if (chatRoomEntryRepository.getAttendanceCount() == MAX_ATTENDANCE_COUNT) {
-            return ChatMessageTranslateResults.to(
-                chatMessageRepository.save(param.toReadChatMessage()),
-                chatMessageRepository.save(param.toReadChatMessage(translatedResponse.translatedText()))
-            );
-        }
-        return ChatMessageTranslateResults.to(
-            chatMessageRepository.save(param.toNotReadChatMessage()),
-            chatMessageRepository.save(param.toNotReadChatMessage(translatedResponse.translatedText()))
-        );
     }
 
     @Transactional
@@ -124,44 +77,6 @@ public class ChatService {
             param.toChatMessageGetRepositoryRequest(CHAT_PAGE_SIZE));
 
         return ChatMessageGetResults.to(chatMessages);
-    }
-
-    private ChatAudioUploadResult uploadAudioFile(ChatAudioUploadParam param) {
-        AudioUploadResponse audioUploadResponse = audioUploader.upload(param.audioFile());
-        ChatMessage chatMessageByStt = chatMessageRepository.save(audioUploadResponse.toChatMessage(param));
-
-        return ChatAudioUploadResult.to(
-            audioUploadResponse.audioUrl(),
-            chatMessageByStt.getCreatedAt()
-        );
-    }
-
-    @Transactional
-    public ChatMessageSttResult sendAudioMessage(ChatMessageSttParam param) {
-        ChatAudioUploadResult chatAudioUploadResult = uploadAudioFile(param.toChatAudioUploadParam());
-
-        SttManager sttManager = sttManagerHandler.getSttManager(SttProvider.GOOGLE.name());
-        SttResponse textBySpeech
-            = sttManager.asyncRecognizeAudio(param.toSttRequest(chatAudioUploadResult.audioUrl()));
-        ChatMessage chatMessage = chatMessageRepository.save(textBySpeech.toChatMessage(param));
-
-        return ChatMessageSttResult.to(
-            textBySpeech.textBySpeech(),
-            chatMessage.getCreatedAt(),
-            SttProvider.GOOGLE.name()
-        );
-    }
-
-    private ChatMessageSttResult fallbackSendAudioMessage(ChatMessageSttParam param, Exception e) {
-        SttManager sttManager = sttManagerHandler.getSttManager(SttProvider.WHISPER.name());
-        SttResponse textBySpeech = sttManager.asyncRecognizeAudio(param.toEmptySttRequest());
-        ChatMessage chatMessage = chatMessageRepository.save(textBySpeech.toChatMessage(param));
-
-        return ChatMessageSttResult.to(
-            textBySpeech.textBySpeech(),
-            chatMessage.getCreatedAt(),
-            SttProvider.WHISPER.name()
-        );
     }
 
     public ChatRoomUserInfoResult getChatRoomUserInfo(ChatRoomUserInfoParam param) {
