+++ src/test/java/com/hellomeritz/global/ControllerTestSupport.java
@@ -1,7 +1,10 @@
 package com.hellomeritz.global;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hellomeritz.chat.controller.ChatController;
 import com.hellomeritz.chat.controller.ChatRoomController;
+import com.hellomeritz.chat.domain.ChatMessage;
+import com.hellomeritz.chat.service.ChatMessageService;
 import com.hellomeritz.chat.service.ChatService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -10,9 +13,10 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
 @WebMvcTest(
-        controllers = {
-                ChatRoomController.class
-        }
+    controllers = {
+        ChatRoomController.class,
+        ChatController.class
+    }
 )
 public abstract class ControllerTestSupport {
 
@@ -27,5 +31,8 @@ public abstract class ControllerTestSupport {
 
     @MockBean
     protected ChatService chatService;
+
+    @MockBean
+    protected ChatMessageService chatMessageService;
 
 }
